[consensus] Update Sui Interface for Mysticeti (#16241)
- Plumb committee/params/protocol config through to AuthorityNode
- Add basic transaction validation
- Plug in mysticeti transaction submission client for Sui
- Plug in consensus output channel to Sui

+++ crates/sui-core/src/consensus_types/mod.rs
@@ -7,4 +7,4 @@ pub(crate) mod consensus_output_api;
 /// An unique integer ID for a validator used by consensus.
 /// In Narwhal, this is the inner value of the `AuthorityIdentifier` type.
 /// In Mysticeti, this is used the same way as the AuthorityIndex type there.
-pub type AuthorityIndex = u16;
+pub type AuthorityIndex = u32;
